+++ src/com/jpexs/games/logik22/AboutDialog.java
@@ -16,7 +16,7 @@ import javax.swing.JPanel;
  */
 public class AboutDialog extends JDialog {
     
-    public static final String VERSION = "1.1.0";
+    public static final String VERSION = "2.0.0";
     public static final String AUTHOR = "JPEXS";
     public static final String YEAR = "2022";
     public static final String VENDOR_ID = "JPEXS";