+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [2.0.0] - 2022-03-02
+- Current field selection (with crosshair)
+- Enter button for processing to next row without remaining colors selection
+- Button for marking field as empty
+- Status bar with hints on mouseover
+
 ## [1.1.0] - 2022-02-27
 - Background color selection
 - Settings are saved to the config file